+++ app/Http/Requests/Diner/StoreDinerSessionRequest.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Requests\Diner;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class StoreDinerSessionRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'guests' => ['required', 'integer', 'min:1'],
+        ];
+    }
+}